Detect build pipeline to use for component (#67)
Signed-off-by: Mykola Morhun mmorhun@redhat.com

What this MR does
This MR adds functionality to choose build pipeline depending on the component. It uses data stored in component's model to make decision. The algorithm of pipeline detection is the following:

Check component's model for language and framework. If they are specified, use available language specific build pipeline.
If the language is not detected, then look for dockerfile and use it
If nothing above works, return noop build as it is unknow how to build the component.